requirements.txt
# Python 3.11+ required for MCP support
openai>=2,<3
pathlib2>=2.3.7
typing-extensions>=4.0.0
google-adk>=1.12.0
tiktoken>=0.7,<1
# Note: google-cloud-storage version is controlled by google-adk dependency
# The FutureWarning about google-cloud-storage <3.0.0 is non-blocking and will be resolved
# when google-adk updates to support google-cloud-storage>=3.0.0